login(redis)/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
redis==4.3.4
Flask-Session==0.4.0
psycopg2-binary==2.9.3
bcrypt==4.0.1